services/cache-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis[hiredis]==5.0.1
aiokafka==0.10.0
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.26.0
aiofiles==23.2.1
